webhook_reciver/requirement.txt
Flask
python-redmine
psycopg2-binary